Add filter by month and refactor redux state
