+++ web/src/pages/paper-trading/actions.js
@@ -1,4 +1,4 @@
-import { addNotification, checkValue } from "../../validations";
+import { addNotification } from "../../validations";
 
 export const GET_TEST_BOTS = "GET_TEST_BOTS";
 export const GET_TEST_BOTS_SUCCESS = "GET_TEST_BOTS_SUCCESS";
@@ -383,44 +383,4 @@ export function saveTestAutoTradeSettings(payload) {
     type: SAVE_TEST_AUTOTRADE_SETTINGS,
     payload: payload,
   };
-}
-
-export function getProfit(base_price, current_price) {
-  if (!checkValue(base_price) && !checkValue(current_price)) {
-    const percent =
-      ((parseFloat(current_price) - parseFloat(base_price)) /
-        parseFloat(base_price)) *
-      100;
-    return percent.toFixed(2);
-  }
-  return 0;
-}
-
-export function computeTotalProfit(bots) {
-  const totalProfit = bots
-    .map((bot) => bot.deal)
-    .reduce((accumulator, currBot) => {
-      let currTotalProfit = getProfit(currBot.buy_price, currBot.current_price);
-      return parseFloat(accumulator) + parseFloat(currTotalProfit);
-    }, 0);
-  return totalProfit.toFixed(2);
-}
-
-export function filterByWeek(bots) {
-  const today = new Date();
-  const lastWeek = new Date(
-    today.getFullYear(),
-    today.getMonth(),
-    today.getDate() - 7
-  );
-  let filteredBots = bots;
-  if (bots.length > 0) {
-    filteredBots = bots.filter((x) => {
-      if (x.created_at) {
-        return x.created_at >= lastWeek.getTime();
-      }
-      return true;
-    });
-  }
-  return filteredBots;
 }
